requirements.txt
pyyaml
guizero
pyserial
